added song files to LofiSpofiy - can play songs in slowmo on app
