+++ Lofi-Spotify/src/app/app.component.html
@@ -1,3 +1,16 @@
-<div>
-  <h3>Lofi Music App</h3>
-</div>
+<!-- <!DOCTYPE html>
+<html lang="en">
+<body>
+    <br/><br/>
+    <button (click)="getSong('Ivy')">Ivy</button>
+    <button (click)="getSong('Gabby')">Gabby</button>
+
+    <!-- <div class="slidecontainer">
+    </div>
+    </div>
+      <input type="range" min="50" max="100" value="25" class="slider" onchange="changeSlide()" id="myRange">
+      <p>Speed: <span id="demo"></span></p>
+    </div> -->
+    
+</body>
+</html> -->
